Removed custom aliases for tmux sessions.
What
===
Removed custom aliases for tmux sessions, replacing them with
instructions for how to use the core tmux commands and setfacl to create
a session, share it with another user, and join it from that user.

Why
===
This devenv should help setup a basic development environment, and then
teach the user how to use tmux for pairing as it is, without creating
custom aliases the user won't find in other environments.