Install Go 1.7.3
What
===
Install Go 1.7.3, setting it up at the system level.

Why
===
The apt-get install golang was installing Go 1.6 which is out of date.
Installing from the main website's download link will keep this image
up-to-date. Additionally moving setup to the installation script and
doing it at the system level rather than at the user level keeps the
setup in a single place, but could still be overriden by the user if
they wished.