Add the `sshenv` alias that sets the SSH env vars.
What
===
Add the `sshenv` alias that sets the SSH Agent environment variables in
the shell that are current in the tmux session.

Why
===
Even though a tmux session gets the latest SSH Agent environment
variables, the shells opened in each pane are not updated. This command
takes the `SSH_` environment variables from the tmux session and exports
them into the current shell.