Remove the system gopath from GOPATH
What
===
Remove the system gopath from GOPATH.

Why
===
While some go tools respect a GOPATH with multiple entries, the guru
tool does not and errors when multiple entries are present. The reason
there are multiple entries in this setup is because tools like guru are
installed in the system GOPATH to remove clutter from the user's GOPATH.
It turns out this is unnecessary. Once the tools are installed there is
no need to keep them in the GOPATH. The only thing that is necessary is
to have the bin directory in the PATH, which is already the case on the
next line.