Fix /usr/local/go/bin not being included in $PATH
What
===
Change the `$GOROOT` reference used when adding the go bin folder to the
path to a new env var `$GO_INSTALL`.

Why
===
In 9e860d2aebde30a750b237f50e11078cde621b87 the `$GOROOT` env var was
removed because @ianlancetaylor mentioned to me at the GopherCon Go
Contributors workshop that it's unnecessary to have this defined, and
in fact because it was defined was intereferring with the source
installation of Go that I was setting up. Unfortunately I neglected to
notice that several lines south I was using `$GOROOT` to include the go
installations bin folder in my `$PATH`.

I don't want to revert the commit because I don't want to reintroduce
`$GOROOT`, so instead introducing a new env var which is not used by go
and just my devenv scripts to identify where go is installed.